Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    int number;
-    scanf("%d",&number);
-    printf("You entered: %d ", number);
+    float number;
+    scanf("%f",&number);
+    printf("You entered: %f ", number);
     return 0;
 }
